ExAequo (ExAequo v0.4.0)

Namespace only and exposing version

Functions

version()

@spec version() :: binary()

Used by the xtra mix task to generate the latest version in the docs, but also handy for client applications for quick exploration in iex.